refactor(domain): improve TransactionForm typing

+++ packages/domain-model/src/models/transactionForm.model.ts
@@ -28,7 +28,7 @@ export const TransactionFormSchema = z
         country: z.optional(z.string().toUpperCase().length(2)),
         vat: z.optional(z.coerce.number().nonnegative().max(1)),
         receipt: z.optional(z.custom<FileWithPath>()),
-        receiptId: z.optional(z.coerce.number()),
+        receiptId: z.optional(z.coerce.number().int()),
     })
     .transform((form) => {
         // Transform amount
